moved new scheduler stuff to core.scheduler package
