+++ bundles/core/org.eclipse.smarthome.core/src/main/java/org/eclipse/smarthome/core/scheduler/CronExpression.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.core.common;
+package org.eclipse.smarthome.core.scheduler;
 
 import java.text.ParseException;
 import java.util.ArrayList;
@@ -16,8 +16,8 @@ import java.util.StringTokenizer;
 import java.util.TimeZone;
 
 import org.apache.commons.lang.StringUtils;
-import org.eclipse.smarthome.core.common.AbstractExpressionPart.BoundedIntegerSet;
-import org.eclipse.smarthome.core.common.CronExpression.CronExpressionPart;
+import org.eclipse.smarthome.core.scheduler.AbstractExpressionPart.BoundedIntegerSet;
+import org.eclipse.smarthome.core.scheduler.CronExpression.CronExpressionPart;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
